refactor: loading, error 페이지 컴포넌트 생성
